fichaxpress-ai/requirements.txt
streamlit
openai
python-dotenv
replicate

